# Add short digest implementation