Create table for generating and looking up short digests.  Any operation involving short codes may be considered secure if the list of digests added to the table is secure.

closes #194