@dmcgowan 2015-03-13
Added benchmark, matches expectations of time complexity...
Add and Dump are linear...Lookup is logarithmic
~~~
BenchmarkAdd10		30000		40637 ns/op
BenchmarkAdd100		3000		458341 ns/op
BenchmarkAdd1000	100  		10566980 ns/op
BenchmarkLookup10	1000000		1369 ns/op
BenchmarkLookup100	1000000		1409 ns/op
BenchmarkLookup1000	1000000		1498 ns/op
BenchmarkDump10		500000		3122 ns/op
BenchmarkDump100	50000		27859 ns/op
BenchmarkDump1000	5000		265343 ns/op
~~~