@stevvooe 2015-03-13
@dmcgowan That `Add` is nasty. Try out a `sort.Search` plus copy-insertion.