@dmcgowan 2015-03-14
Added commit which uses search + slice copy, as expected it is faster for larger arrays

~~~
benchmark               old ns/op     new ns/op     delta
BenchmarkAdd10          40884         43642         +6.75%
BenchmarkAdd100         472603        430000        -9.01%
BenchmarkAdd1000        10493745      4576972       -56.38%
~~~