@stevvooe 2015-03-21
Let's call the main object `ShortTable` a `DigestSet` or something like that. Let's remove the `DumpTable` for now, since I'm a little shaky on the use case.

Do we also want to be able to iterate over a `DigestSet`?