@dmcgowan 2015-03-23
Renamed to DigestSet and cleaned up some of the comments to explain its general use and relevance to short string lookups.  Also moved the dump table function to a non-member function which takes in a digest.

The name `DigestSet` does stutter and I would prefer to pick a name which is more lint friendly. The name `Set` may be more appropriate given its generalized use case. WDYT?