@stevvooe 2015-03-23
@dmcgowan I am fine going with `Set`, since, as we discussed, this very accurate.