@dmcgowan 2015-03-24
Updated to `Set` and went ahead and squashed to one clean commit